Upstream commits 581f08f..7f63cee
Change third party srl branch

Speed up ray-casting

Fix node sample point in projective functor

Refactor projective functor

- Solve non thread save problem
- Accelerate projection for PinholeCamera without distortion parameters

Add enable print flag to reader to suppress warnings

Add option to save octree structure

Save octree structure when mesh_volume or the final frame is true

Add option to access the number of allocated nodes, blocks and scales

Save the number of nodes, blocks and scales to result file

Add number of nodes, blocks and scales table to summary

Add enable structure to benchmark

Add number of nodes, blocks and scales to SEStats

Plot number of nodes, blocks and scales

Refactor get function

- Remove bugged get(...) function
- Rename getFine(...) functions to get(...)
- Return scale the data is returned from
  - Pass data as argument

Scale meshing source and destination offset by the voxel size

Implement feature to save 3d slice with integration scale info

Add VoxelBlockFinest voxel block class

Save scale slice

Add selection and checking function to voxel implementations

Refactor fetch and fetchNode

Clean up idw interpolation

Clean up fetch neighbour functions

Clean up save 3d slice function

Add save 3d slice with different  selection functions for node and voxel

Use save 3d slice functions with different select functions

Add find valid point function for one select value function

Use voxel implementation select and is inside check for meshing

Clean up raycasting and rendering kernels

Use voxel implementation select and checking functions for rendering

Refactor gradient function

Fix compiler warnings

Fix the power monitor not returning power usage

Accidentally removed a line in a previous commit.

Fix PowerMonitor.cpp indentation

Fix installation issues

Fixup - Raycasting

Verify that the peek point is inside the map

Fix assignments to references

Fix valid point test

Add median to math utils

Use median downsampling for depth images

Move performance stats

Refactor performance stats

New performance stats

Apply new performance stats to python scripts

Rename and namespace static perfstats variable

Update README to new API

Set srl float type in hpp file rather than cmake

Remove Sophus

Fix PinholeCamera intrinsics downsampling

Add googletest as a submodule

Move voxel impl config to DenseSLAM

Add FindEigen3.cmake

Move se::Volume into separate header

Use VoxelType::isValid in the octree iterator

Fix missing const in Octree::contains()

Add Node function to get own data

Add FOV computation in sensor

Add a median that does not create values not in the original data

Issue found by Nils.

Remove specific Python minor version from script

Simplify the arguments to result_summary.py

No need to do
  result_summary.py --result-file input.txt --summary-file output
when we can do
  result_summary.py input.txt output

Fix calling the wrong grad()

Found by clang.

Fix warnings reported by clang

Fix undefined behavior and potential memory leak

Fix compilation error when using clang

No idea why it fails.

Fix error in operator order

The expression computed before was (!(x) == 0) instead of !(x == 0).
Found by clang.

Fix cast causing loss of precision

2^31 - 1 (the max int on most 64bit architectures) can't be represented
accurately in a 32bit float.

Fix VoxelImpl parameter not being set

Explicitly link some libraries when using clang

Fix googletest submodule to use HTTPS

+++ se_apps/src/reader_iclnuim.cpp
@@ -124,8 +124,8 @@ se::ReaderStatus se::ICLNUIMReader::nextRGBA(se::Image<uint32_t>& rgba_image) {
     free(image_data);
     return se::ReaderStatus::error;
   }
-  assert(rgba_image_res_.x() == w);
-  assert(rgba_image_res_.y() == h);
+  assert(rgba_image_res_.x() == static_cast<int>(w));
+  assert(rgba_image_res_.y() == static_cast<int>(h));
   // Resize the output image if needed.
   if ((   rgba_image.width()  != rgba_image_res_.x())
       || (rgba_image.height() != rgba_image_res_.y())) {
